Added new tests
